pyproject.toml
[project]
name = "feather-etl"
version = "0.1.0"
description = "Config-driven ETL package for extracting data from heterogeneous sources into local DuckDB"
readme = "README.md"
authors = [
    { name = "Siraj Samsudeen", email = "mailsiraj@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.0",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "pyodbc>=5.0",
    "apscheduler>=3.10",
    "openpyxl>=3.1",
    "pytz",
]

[project.scripts]
feather = "feather.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/feather"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
